+++ app/models/dashboard_role.rb
@@ -24,7 +24,4 @@ class DashboardRole < ActiveRecord::Base
 
   belongs_to :dashboard
   belongs_to :role
-
-  validates :dashboard, :role,
-            presence: true
 end