+++ test/unit/blocks/calendar_block_test.rb
@@ -1,10 +1,6 @@
-<%
 # frozen_string_literal: true
 
 # This file is part of the Plugin Redmine Dashboards.
-#
-# Copyright (C) 2016 - 2021 Alexander Meindl <https://github.com/alexandermeindl>, alphanodes.
-# See <https://github.com/AlphaNodes/additionals>.
 #
 # Copyright (C) 2021 - 2022 Liane Hampe <liaham@xmera.de>, xmera.
 #
@@ -21,6 +17,19 @@
 # You should have received a copy of the GNU General Public License
 # along with this program; if not, write to the Free Software
 # Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
-%>
 
-$("#block-<%= block_id %>").replaceWith("<%= escape_javascript show(dashboard).render_dashboard_block(block_id.to_s, sort_options) %>");
+require File.expand_path '../../test_helper', __dir__
+
+class CalendarBlockTest < RedmineDashboards::TestCase
+  def setup
+    @calendar_block = CalendarBlock.instance
+  end
+
+  def teardown
+    @calendar_block = nil
+  end
+
+  def test_respond_to_calendar
+    assert @calendar_block.respond_to? :calendar
+  end
+end