+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.1.0 - 2022-09-21
+
+### Added
+
+* calendar block
+* spent time block with entries
+* open approval block of redmine_dmsf
+* locked files block of redmine_dmsf
+* checkbox for disabling My Page
+
+### Changes
+
+* BasePresenter class by integrating AdvancedPluginHelper
+
 ## 1.0.4 - 2022-06-29
 
 ### Changes